requirements.txt
commonutils==0.1.0
ipython==7.30.0
joblib==1.1.0
matplotlib==3.5.0
modin==0.12.0
networkx==2.6.3
numpy==1.19.5
pandas==1.3.4
psutil==5.8.0
pyspark==3.2.0
PyYAML==6.0
ray==1.9.0
scikit_learn==1.0.2
scipy==1.7.1
seaborn==0.11.2
six==1.15.0
statsmodels==0.12.1
swifter==1.0.9
tensorflow_cpu==2.7.0
tensorflow_addons==0.16.1
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
typeguard==2.13.3
wget==3.2
